Add coverage checker, random word generator.
This should likely be split into a couple of files, but for
now I've added code for a coverage check. You'd have to tweak main
to run it.

Also, because so many words aren't covered, I added a method that
generates a random word-like sequence if you ask for a word that
doesn't exist.