Reshape repo, add romaji-converter
In order to add some new data sources, I need a utility to convert romaji to kana.

I wrote the start of a romaji normalizer (there are many ways to romanize
Japanese so it would be nice to somewhat unify) and to adding a test made
me feel like I should rework the structure of my code a bit.

Then I added a romaji->hiragana method. This should be useful for my data
ingest.