Add parsing of jmnedict data. Update coverage checker.
This adds parsing for the three jmnedict file types.

This did great stuff for coverage of small number lengths.

Some Coverage Records:
1 digits: 100%
2 digits: 100%
3 digits: ~88%~ -> ~96.80%~ -> 100%
4 digits: ~68%~ -> ~86.84%~ -> 99.98% [missing only 9585,9588]
5 digits: 65.67% -> 98.83%
6 digits: 36.47% -> 91.13%
7 digits: 11.13% -> 65.39%
8 digits: 22.96% [this took around 3 minutes to run]
9 digits: 2.88% [Also this took nearly half an hour]

In order to run this coverage info, I had to add some helpers and I'm not
pleased with the state of some of the code currently. A lot of copy and paste
has started spreading.

This also adds methods that get _any_ example instead of _all_ examples. Now that
there is so much more data, the _all_ example version was starting to chug in coverage.

This was a good pick because it's still pretty slow all things considered. I didn't even
try to check coverage of 10 digits because 9 digits, even after stripping out all the
printing, took nearly 30 minutes.